# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "### _API for student info collection service._",
        "title": "GolangUnitedSchool",
        "contact": {},
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/homeworks": {
            "get": {
                "description": "возвращает все домашние работы",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "homeworks"
                ],
                "summary": "get all items in the homework list",
                "operationId": "get-all-homeworks",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.HomeworkList"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.ResponseMessage"
                        }
                    }
                }
            },
            "post": {
                "description": "добавляет новую домашнюю работу",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "homeworks"
                ],
                "summary": "add new homework to the homework list",
                "operationId": "create-homework",
                "parameters": [
                    {
                        "description": "homework",
                        "name": "homework",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Homework"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.ResponseMessage"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.ResponseMessage"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.ResponseMessage"
                        }
                    }
                }
            }
        },
        "/homeworks/{id}": {
            "get": {
                "description": "возвращает домашнюю работу с указанным id",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "homeworks"
                ],
                "summary": "get a homework by ID",
                "operationId": "get-homework-by-id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "homework id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Homework"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/model.ResponseMessage"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.ResponseMessage"
                        }
                    }
                }
            },
            "put": {
                "description": "изменяет домашнюю работу с указанным id",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "homeworks"
                ],
                "summary": "update a homework by ID",
                "operationId": "update-homework-by-id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "homework id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "homework",
                        "name": "homework",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Homework"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.ResponseMessage"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.ResponseMessage"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/model.ResponseMessage"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.ResponseMessage"
                        }
                    }
                }
            },
            "delete": {
                "description": "удаляет домашнюю работу с указанным id",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "homeworks"
                ],
                "summary": "delete a homework by ID",
                "operationId": "delete-homework-by-id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "homework id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.ResponseMessage"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/model.ResponseMessage"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.ResponseMessage"
                        }
                    }
                }
            }
        },
        "/interview": {
            "get": {
                "description": "возвращает все интервью",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "interviews"
                ],
                "summary": "get all items in the interview list",
                "operationId": "get-all-interviews",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.Interview"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.ResponseMessage"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.ResponseMessage"
                        }
                    }
                }
            },
            "post": {
                "description": "добавляет новое интервью новое интервью",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "interviews"
                ],
                "summary": "add new item to the interview list",
                "operationId": "create-interview",
                "parameters": [
                    {
                        "description": "interview",
                        "name": "interview",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Interview"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/model.ResponseMessage"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.ResponseMessage"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.ResponseMessage"
                        }
                    }
                }
            }
        },
        "/interview/{id}": {
            "get": {
                "description": "возвращает интервью с указанным id",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "interviews"
                ],
                "summary": "get a interview by ID",
                "operationId": "get-interview-by-id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "interview id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Interview"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/model.ResponseMessage"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.ResponseMessage"
                        }
                    }
                }
            },
            "put": {
                "description": "изменяет интервью с указанным id",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "interviews"
                ],
                "summary": "put a interview by ID",
                "operationId": "update-interview-by-id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "interview id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "interview",
                        "name": "interview",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Interview"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.ResponseMessage"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.ResponseMessage"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/model.ResponseMessage"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.ResponseMessage"
                        }
                    }
                }
            },
            "delete": {
                "description": "удаляет интервью с указанным id",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "interviews"
                ],
                "summary": "delete a interview by ID",
                "operationId": "delete-interview-by-id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "interview id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.ResponseMessage"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/model.ResponseMessage"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.ResponseMessage"
                        }
                    }
                }
            }
        },
...
        "/lectures/{lecture_id}/homeworks": {
            "get": {
                "description": "возвращает домашнюю работы по указанному id лекции",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "homeworks"
                ],
                "summary": "get items in the homework list by lecture ID",
                "operationId": "get-homeworks-by-lecture-id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "lecture id",
                        "name": "lecture_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.HomeworkList"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/model.ResponseMessage"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.ResponseMessage"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "model.Homework": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "lecture_id": {
                    "type": "integer"
                },
                "max_score": {
                    "type": "number"
                },
                "task": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "model.HomeworkList": {
            "type": "object",
            "properties": {
                "_metadata": {
                    "$ref": "#/definitions/model.PaginationResponse"
                },
                "homeworks": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Homework"
                    }
                }
            }
        },
        "model.Interview": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "mentorID": {
                    "type": "integer"
                },
                "note": {
                    "type": "string"
                },
                "score": {
                    "type": "integer"
                },
                "studentID": {
                    "type": "integer"
                }
            }
        },
...
        "model.PaginationResponse": {
            "type": "object",
            "properties": {
                "page": {
                    "type": "integer"
                },
                "page_count": {
                    "type": "integer"
                },
                "page_size": {
                    "type": "integer"
                },
                "total_count": {
                    "type": "integer"
                }
            }
        },
...
        "model.ResponseMessage": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
    }
}